- Changed the script to take a third command line param as filename for   the different books - Fixed some problems with empty screenshots directory

+++ scripts/auto_build/docbuild.sh
@@ -1,9 +1,9 @@
 #!/bin/sh
 # --
 # scripts/auto_build/docbuild.sh - Automated creation of the  OTRS docu
-# Copyright (C) 2001-2006 OTRS GmbH, http://otrs.org/
+# Copyright (C) 2001-2007 OTRS GmbH, http://otrs.org/
 # --
-# $Id: docbuild.sh,v 1.7 2006-10-31 15:43:24 tr Exp $
+# $Id: docbuild.sh,v 1.8 2007-02-20 15:56:35 cs Exp $
 # --
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -30,14 +30,25 @@ TMP="/tmp"
 PACKAGE_BUILD_DIR="$TMP/$PACKAGE-build"
 PACKAGE_DEST_DIR="$TMP/$PACKAGE-package"
 
+if [ $# -ne 3 ]; then
+    # --
+    # build src needed
+    # --
+    echo ""
+    echo "Usage: docbuild.sh <PATH_TO_DOC_SRC> <PATH_TO_OTRS_FRAMEWORK_SRC> <NAME"
+    echo ""
+    echo "  Try: docbuild.sh /home/ernie/src/doc-admin /home/ernie/src/otrs-cvs otrs_admin_book"
+    echo ""
+    exit 1;
+fi
 if ! test -e $PATH_TO_CVS_SRC; then
     # --
     # build src needed
     # --
     echo ""
-    echo "Usage: docbuild.sh <PATH_TO_CVS_SRC> <PATH_TO_CVS_FRAMEWORK_SRC>"
+    echo "Usage: docbuild.sh <PATH_TO_DOC_SRC> <PATH_TO_OTRS_FRAMEWORK_SRC> <NAME"
     echo ""
-    echo "  Try: docbuild.sh /home/ernie/src/otrs-doc /home/ernie/src/otrs-cvs"
+    echo "  Try: docbuild.sh /home/ernie/src/doc-admin /home/ernie/src/otrs-cvs otrs_admin_book"
     echo ""
     exit 1;
 fi
@@ -46,14 +57,14 @@ if ! test -e $PATH_TO_CVS_FRAMEWORK_SRC; then
     # build src needed
     # --
     echo ""
-    echo "Usage: docbuild.sh <PATH_TO_CVS_SRC> <PATH_TO_CVS_FRAMEWORK_SRC>"
+    echo "Usage: docbuild.sh <PATH_TO_DOC_SRC> <PATH_TO_CVS_FRAMEWORK_SRC>"
     echo ""
-    echo "  Try: docbuild.sh /home/ernie/src/otrs-doc /home/ernie/src/otrs-cvs"
+    echo "  Try: docbuild.sh /home/ernie/src/doc-admin /home/ernie/src/otrs-cvs otrs_admin_book"
     echo ""
     exit 1;
 fi
 # check dir
-if ! test -e $PATH_TO_CVS_SRC/de; then
+if ! test -e $PATH_TO_CVS_SRC/en; then
     echo "Error: $PATH_TO_CVS_SRC is not OTRS CVS directory!"
     exit 1;
 fi
@@ -64,6 +75,10 @@ mkdir -p $PACKAGE_DEST_DIR/
 
 
 for Language in en de; do
+    if ! test -e $PATH_TO_CVS_SRC/$Language; then
+        continue
+    fi
+
     # prepare build env
     rm -rf $PACKAGE_BUILD_DIR || exit 1;
     mkdir -p $PACKAGE_BUILD_DIR/ || exit 1;
@@ -83,35 +98,33 @@ for Language in en de; do
     $PATH_TO_CVS_FRAMEWORK_SRC/scripts/tools/charset-convert.pl -s utf-8 -d iso-8859-1 -f $PACKAGE_BUILD_DIR/$Language/all-config-parameters.xml
 
     # create one big xml file with all chapters
-    xmllint --xinclude book.xml > otrs_admin_book.xml
+    xmllint --xinclude book.xml > $3.xml
 
     # pdf
-    docbook2pdf otrs_admin_book.xml
+    docbook2pdf $3.xml
     mkdir -p $PACKAGE_DEST_DIR/$Language/pdf
-    cp otrs_admin_book.pdf $PACKAGE_DEST_DIR/$Language/pdf/otrs_admin_book.pdf
+    cp $3.pdf $PACKAGE_DEST_DIR/$Language/pdf/$3.pdf
 
     # txt
-    docbook2txt otrs_admin_book.xml
+    docbook2txt $3.xml
     mkdir -p $PACKAGE_DEST_DIR/$Language/txt
-    cp otrs_admin_book.txt $PACKAGE_DEST_DIR/$Language/txt/otrs_admin_book.txt
+    cp $3.txt $PACKAGE_DEST_DIR/$Language/txt/$3.txt
 
     # html
-    docbook2html otrs_admin_book.xml
+    docbook2html $3.xml
     mkdir -p $PACKAGE_DEST_DIR/$Language/html
-    mkdir -p $PACKAGE_DEST_DIR/$Language/html/screenshots
-    mkdir -p $PACKAGE_DEST_DIR/$Language/images
     cp -R *.html $PACKAGE_DEST_DIR/$Language/html/
-    cp -R screenshots/* $PACKAGE_DEST_DIR/$Language/html/screenshots/
-    cp -R images/* $PACKAGE_DEST_DIR/$Language/images/
+    cp -R screenshots $PACKAGE_DEST_DIR/$Language/html/
+    cp -R images $PACKAGE_DEST_DIR/$Language/
 
     # test for convert included in imagemagick package
     if [ -x /usr/bin/convert ] || [ -x /bin/convert ] ; then
         # convert images to 60% of orig. size
         CONVERT=`which convert`
-        for i in $PACKAGE_DEST_DIR/$Language/html/screenshots/*.png ; do
+        for i in `ls -1 $PACKAGE_DEST_DIR/$Language/html/screenshots/*.png 2>/dev/null`; do
             echo "convert image to 60% $i"
-        $CONVERT $i -resize 60% $i
-    done;
+            $CONVERT $i -resize 60% $i
+        done;
     fi
 
     # xml
@@ -128,12 +141,15 @@ echo ""
 echo "Builded packages and files:"
 echo "---------------------------"
 for Language in en de; do
+    if ! test -e $PACKAGE_DEST_DIR/$Language; then
+        continue
+    fi
     du -sh $PACKAGE_DEST_DIR/$Language/xml/
     du -sh $PACKAGE_DEST_DIR/$Language/html/
     du -sh $PACKAGE_DEST_DIR/$Language/pdf/
     du -sh $PACKAGE_DEST_DIR/$Language/txt/
     echo ""
-    ls -l $PACKAGE_DEST_DIR/$Language/pdf/otrs_admin_book.pdf;
-    ls -l $PACKAGE_DEST_DIR/$Language/txt/otrs_admin_book.txt;
+    ls -l $PACKAGE_DEST_DIR/$Language/pdf/$3.pdf;
+    ls -l $PACKAGE_DEST_DIR/$Language/txt/$3.txt;
     echo ""
 done;
